+++ magda-web-client/src/AppContainer.js
@@ -1,6 +1,6 @@
 import MagdaDocumentTitle from "./Components/i18n/MagdaDocumentTitle";
 import React from "react";
-import { config } from "./config.js";
+import { config } from "./config";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 
@@ -19,7 +19,7 @@ import { Route, Switch } from "react-router-dom";
 import { Medium } from "./UI/Responsive";
 import AUskipLink from "./pancake/react/skip-link";
 
-import "./AppContainer.css";
+import "./AppContainer.scss";
 
 class AppContainer extends React.Component {
     componentDidMount() {