+++ magda-web-client/src/Components/Footer/Feedback/FeedbackForm.js
@@ -11,7 +11,7 @@ import close from "../../../assets/close.svg";
 import success from "../../../assets/success.svg";
 import Notification from "../../../UI/Notification";
 
-import "./FeedbackForm.css";
+import "./FeedbackForm.scss";
 
 class FeedbackForm extends React.Component {
     constructor(props) {