+++ magda-web-client/src/helpers/defined.js
@@ -1,4 +1,0 @@
-// @flow
-export default function defined(value: any): boolean {
-    return value !== undefined && value !== null;
-}